Refactor 'elements' into layouts and widgets
